refactor: move api spec code to `mithril-api-spec` crate
