+++ mithril-aggregator/Cargo.toml
@@ -77,6 +77,7 @@ tikv-jemallocator = { version = "0.6.0", optional = true }
 criterion = { version = "0.6.0", features = ["html_reports", "async_tokio"] }
 http = "1.3.1"
 httpmock = "0.7.0"
+mithril-api-spec = { path = "../internal/tests/mithril-api-spec" }
 mithril-common = { path = "../mithril-common", features = [
     "allow_skip_signer_certification",
     "test_tools",
