+++ mithril-aggregator/src/http_server/routes/artifact_routes/cardano_database.rs
@@ -176,21 +176,24 @@ mod handlers {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::{initialize_dependencies, services::MockMessageService};
-    use mithril_common::messages::{
-        CardanoDatabaseDigestListItemMessage, CardanoDatabaseSnapshotListItemMessage,
-        CardanoDatabaseSnapshotMessage,
-    };
-    use mithril_common::test_utils::apispec::APISpec;
-    use mithril_common::{MITHRIL_CLIENT_TYPE_HEADER, MITHRIL_ORIGIN_TAG_HEADER};
-    use mithril_persistence::sqlite::HydrationError;
     use serde_json::Value::Null;
     use std::sync::Arc;
     use warp::{
         http::{Method, StatusCode},
         test::request,
     };
+
+    use mithril_api_spec::APISpec;
+    use mithril_common::messages::{
+        CardanoDatabaseDigestListItemMessage, CardanoDatabaseSnapshotListItemMessage,
+        CardanoDatabaseSnapshotMessage,
+    };
+    use mithril_common::{MITHRIL_CLIENT_TYPE_HEADER, MITHRIL_ORIGIN_TAG_HEADER};
+    use mithril_persistence::sqlite::HydrationError;
+
+    use crate::{initialize_dependencies, services::MockMessageService};
+
+    use super::*;
 
     fn setup_router(
         state: RouterState,
