+++ mithril-aggregator/src/http_server/routes/certificate_routes.rs
@@ -94,15 +94,16 @@ mod handlers {
 #[cfg(test)]
 mod tests {
     use anyhow::anyhow;
-    use mithril_common::{
-        test_utils::{apispec::APISpec, fake_data},
-        MITHRIL_CLIENT_TYPE_HEADER, MITHRIL_ORIGIN_TAG_HEADER,
-    };
     use serde_json::Value::Null;
     use std::sync::Arc;
     use warp::{
         http::{Method, StatusCode},
         test::request,
+    };
+
+    use mithril_api_spec::APISpec;
+    use mithril_common::{
+        test_utils::fake_data, MITHRIL_CLIENT_TYPE_HEADER, MITHRIL_ORIGIN_TAG_HEADER,
     };
 
     use crate::{initialize_dependencies, services::MockMessageService};
