+++ mithril-common/src/lib.rs
@@ -60,21 +60,3 @@ pub const MITHRIL_CLIENT_TYPE_HEADER: &str = "mithril-client-type";
 macro_rules! era_deprecate {
     ( $comment:literal ) => {};
 }
-
-#[cfg(test)]
-mod tests {
-    #[cfg(feature = "apispec")]
-    #[test]
-    fn test_openapi_examples_conformity() {
-        use crate::test_utils::apispec::APISpec;
-        let api_spec = APISpec::from_file(crate::test_utils::apispec::DEFAULT_SPEC_FILE);
-
-        let errors: Vec<String> = api_spec.verify_examples();
-
-        assert!(
-            errors.is_empty(),
-            "Errors in examples\n{}",
-            errors.join("\n")
-        );
-    }
-}
